Promote text display mixin to Module Nav instead of Accordion
